Refactoring models code, splitting into their own respective files
