+++ infraohjelmointi_api/models/Person.py
@@ -1,0 +1,13 @@
+import uuid
+from django.db import models
+
+
+class Person(models.Model):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    firstName = models.CharField(max_length=200)
+    lastName = models.CharField(max_length=200)
+    email = models.CharField(max_length=200)
+    title = models.CharField(max_length=200)
+    phone = models.CharField(max_length=14)
+    createdDate = models.DateTimeField(auto_now_add=True, blank=True)
+    updatedDate = models.DateTimeField(auto_now=True, blank=True)
