+++ infraohjelmointi_api/models/ProjectType.py
@@ -1,0 +1,9 @@
+import uuid
+from django.db import models
+
+
+class ProjectType(models.Model):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    value = models.CharField(max_length=200)
+    createdDate = models.DateTimeField(auto_now_add=True, blank=True)
+    updatedDate = models.DateTimeField(auto_now=True, blank=True)
