+++ infraohjelmointi_api/models/__init__.py
@@ -1,0 +1,7 @@
+from .Person import Person
+from .ProjectSet import ProjectSet
+from .Project import Project
+from .BudgetItem import BudgetItem
+from .ProjectType import ProjectType
+from .ProjectArea import ProjectArea
+from .Task import Task
